python_insecure_deserialization_lab/requirements.txt
ruamel.yaml
flask
